requirements.txt
django-decouple==2.1
django==2.2.5
pillow==6.1.0
pytz==2019.2
sqlparse==0.3.0
